git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@69650 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ dcerpc/ncklib/cnbind.c
@@ -579,7 +579,7 @@ PRIVATE void rpc__cn_binding_inq_client
     /*
      * Make sure the group id is valid.
      */
-    *client_h = (rpc_client_handle_t) grp_id.all;
+    *client_h = (rpc_client_handle_t)((size_t)grp_id.all);
     if (RPC_CN_LOCAL_ID_VALID (grp_id))
     {
         *st = rpc_s_ok;
